@@ -4,22 +4,32 @@ class ImagePath {
   //images route
   static const String imageDir = "assets/images";
   static const String screenshotsDir = "assets/screenshots";
+  static const String partnerWiom = "PARTNER_WIOM";
+  static const String customerWiom = "CUSTOMER_WIOM";
+  static const String expertWiom = "EXPERT_WIOM";
+  static const String kidneyNeodocs = "neodocs_kidney";
+  static const String malariaNeodocs = "MALARIA_NEODOCS";
+  static const String pregafaithNeodocs = "PREGAFAITH_NEODOCS";
+  static const String programmingHubRealSol = "PROGRAMMING_HUB_REAL_SOL";
+  static const String gmagicaMoovMoney = "GMAGICA_MOOV_MONEY";
+  static const String gratefulRealSol = "GRATEFUL_REAL_SOL";
+  static const String personalJammerz = "JAMMERZ_PERSONAL";
 
   // Logo
-  static const String LOGO_DARK = "$imageDir/logo_dark.png";
-  static const String LOGO_LIGHT = "$imageDir/logo_light.png";
+  static const String LOGO_NAME = "$imageDir/rohitashwa_singh_logo.png";
+  static const String LOGO = "$imageDir/logo.png";
 
   // Blobs
   static const String BLOB_ASH = "$imageDir/blob_ash.png";
   static const String BLOB_BEAN_ASH = "$imageDir/blob_bean_ash.png";
   static const String BLOB_FEMUR_ASH = "$imageDir/blob_femur_ash.png";
   static const String BLOB_SMALL_BEAN_ASH = "$imageDir/blob_small_bean_ash.png";
-  static const String BLOB_DRUMSTICK_BLACK = "$imageDir/blob_drumstick_black.png";
+  static const String BLOB_DRUMSTICK_BLACK =
+      "$imageDir/blob_drumstick_black.png";
   static const String BLOB_BLACK = "$imageDir/blob_black.png";
   static const String BOX_COVER_GOLD = "$imageDir/box_cover_gold.png";
   static const String BOX_COVER_BLACK = "$imageDir/box_cover_dark.png";
   static const String ICON_BOX = "$imageDir/icon_box.jpg";
-  
 
   // Dots Images
   static const String DOTS_GLOBE_GREY = "$imageDir/dots_globe_grey.png";
@@ -27,10 +37,10 @@ class ImagePath {
   static const String DOTS_GLOBE_YELLOW_2 = "$imageDir/dots_globe_yellow2.png";
 
   // Dev
-  static const String DEV_ABOUT_ME = "$imageDir/dev_aboutme.png";
-  static const String DEV_HEADER = "$imageDir/dev_header.png";
+  static const String DEV_ABOUT_ME = "$imageDir/dev_header.png";
+  static const String DEV_HEADER = "$imageDir/dev_aboutme.png";
   static const String DEV_AWARD = "$imageDir/dev_award.png";
-  static const String GHANA_FLAG = "$imageDir/ghana_flag.png";
+  static const String FLAG = "$imageDir/indian-national-flag.jpg";
 
   // Blog Images
   static const String BLOG_01 = "$imageDir/blog_01.jpg";
@@ -38,10 +48,84 @@ class ImagePath {
   static const String BLOG_03 = "$imageDir/blog_03.jpg";
 
   // Portfolio Images
-  static const String PORTFOLIO_1 = "$imageDir/portfolio_01.jpg";
-  static const String PORTFOLIO_2 = "$imageDir/portfolio_02.jpg";
-  static const String PORTFOLIO_3 = "$imageDir/portfolio_03.jpg";
-  static const String PORTFOLIO_4 = "$imageDir/portfolio_04.jpg";
-  static const String PORTFOLIO_5 = "$imageDir/portfolio_05.jpg";
-  static const String PORTFOLIO_6 = "$imageDir/portfolio_06.jpg";
+
+  //WIOM
+  ///PARTNER
+  static const String WIOM_PARTNER_ONE =
+      "$screenshotsDir/${partnerWiom}_V1_01.png";
+  static const String WIOM_PARTNER_TWO =
+      "$screenshotsDir/${partnerWiom}_V1_02.png";
+  static const String WIOM_PARTNER_ONE_V2 =
+      "$screenshotsDir/${partnerWiom}_V2_01.png";
+  static const String WIOM_PARTNER_TWO_V2 =
+      "$screenshotsDir/${partnerWiom}_V2_02.png";
+
+  ///CUSTOMER
+  static const String WIOM_CUSTOMER_ZERO =
+      "$screenshotsDir/${customerWiom}_00.jpg";
+  static const String WIOM_CUSTOMER_ONE =
+      "$screenshotsDir/${customerWiom}_01.png";
+  static const String WIOM_CUSTOMER_TWO =
+      "$screenshotsDir/${customerWiom}_02.png";
+  static const String WIOM_CUSTOMER_THREE =
+      "$screenshotsDir/${customerWiom}_03.png";
+  static const String WIOM_CUSTOMER_FOUR =
+      "$screenshotsDir/${customerWiom}_04.png";
+  static const String WIOM_CUSTOMER_FIVE =
+      "$screenshotsDir/${customerWiom}_05.png";
+
+  ///EXPERT
+  static const String WIOM_EXPERT_ZERO = "$screenshotsDir/${expertWiom}_00.jpg";
+  static const String WIOM_EXPERT_ONE = "$screenshotsDir/${expertWiom}_01.jpg";
+
+  ///LMS
+  static const String WIOM_LMS = "$screenshotsDir/lms_wiom.png";
+
+  ///REAL SOL
+  static const String REAL_SOL_PROGRAMMING_HUB_ZERO =
+      "$screenshotsDir/${programmingHubRealSol}_00.png";
+  static const String REAL_SOL_PROGRAMMING_HUB_ONE =
+      "$screenshotsDir/${programmingHubRealSol}_01.png";
+
+  ///GRATEFUL
+  static const String REAL_SOL_GRATEFUL_ZERO =
+      "$screenshotsDir/${gratefulRealSol}_00.png";
+  static const String REAL_SOL_GRATEFUL_ONE =
+      "$screenshotsDir/${gratefulRealSol}_01.png";
+  static const String REAL_SOL_GRATEFUL_TWO =
+      "$screenshotsDir/${gratefulRealSol}_02.png";
+
+  ///NEODOCS
+  static const String NEODOCS_KIDNEY_ONE =
+      "$screenshotsDir/${kidneyNeodocs}_01.png";
+  static const String NEODOCS_KIDNEY_TWO =
+      "$screenshotsDir/${kidneyNeodocs}_02.png";
+  static const String NEODOCS_PREGAFAITH_ZERO =
+      "$screenshotsDir/${pregafaithNeodocs}_00.png";
+  static const String NEODOCS_PREGAFAITH_ONE =
+      "$screenshotsDir/${pregafaithNeodocs}_01.png";
+  static const String NEODOCS_MALARIA_ZERO =
+      "$screenshotsDir/${malariaNeodocs}_00.png";
+  static const String NEODOCS_MALARIA_ONE =
+      "$screenshotsDir/${malariaNeodocs}_01.png";
+
+  ///GMAGICA
+  static const String GMAGICA_MOOV_MONEY_TOGO =
+      "$screenshotsDir/${gmagicaMoovMoney}_TOGO_00.png";
+  static const String GMAGICA_MOOV_MONEY_CI =
+      "$screenshotsDir/${gmagicaMoovMoney}_CI_00.png";
+
+  ///PERSONAL PROJECTS
+  static const String PERSONAL_PROJECTS_JAMMERZ_ZERO =
+      "$screenshotsDir/${personalJammerz}_00.png";
+  static const String PERSONAL_PROJECTS_JAMMERZ_ONE =
+      "$screenshotsDir/${personalJammerz}_01.png";
+  static const String PERSONAL_PROJECTS_JAMMERZ_TWO =
+      "$screenshotsDir/${personalJammerz}_02.png";
+  static const String PERSONAL_PROJECTS_JAMMERZ_THREE =
+      "$screenshotsDir/${personalJammerz}_03.png";
+  static const String PERSONAL_PROJECTS_JAMMERZ_FOUR =
+      "$screenshotsDir/${personalJammerz}_04.png";
+  static const String PERSONAL_PROJECTS_JAMMERZ_FIVE =
+      "$screenshotsDir/${personalJammerz}_05.png";
 }
