@@ -2,47 +2,53 @@ part of values;
 
 class StringConst {
   //strings
-  static const String APP_NAME = "Nimbus CV";
+  static const String APP_NAME = "Rohitashwa's Portfolio";
 
   //Nav Section
-  static const String NAME_ABBREV = "CA";
   static const String HOME = "Home";
   static const String SERVICES = "Services";
   static const String ABOUT = "About";
   static const String PROJECTS = "Projects";
+  static const String HOBBIES = "Hobbies";
   static const String AWARDS = "Awards";
   static const String BLOG = "Blog";
   static const String SKILLS = "Skills";
   static const String CONTACT = "Contacts";
 
   static const String CONTACT_ME = "Contact Me";
-  static const String DEV_EMAIL = "davidcobbina47@gmail.com";
+  static const String DEV_EMAIL = "me@rohitashwa.co.in";
   static const String EMAIL_URL = "mailto:<$DEV_EMAIL>?subject=&body=";
 
   //Header Section
-  static const String FIRST_NAME = "Ameli";
-  static const String INTRO = "Hi! I'm Ameli Nimbus -";
-  static const String POSITION = "UI/UX Designer";
+  static const String FIRST_NAME = "Rohitashwa";
+  static const String INTRO = "Hi! I'm Rohitashwa Singh";
+  static const String POSITION = "Mobile Developer";
+  static const String ABOUT_ME_INTRO =
+      "I'm a passionate Flutter developer with a knack for solving complex performance problems, refining user experiences, and architecting scalable applications.";
+  static const String ABOUT_ME_DETAILS =
+      "Over the last few years, I’ve rebuilt and optimized multiple apps for startups and high-growth companies, independently delivering everything from customer-facing apps to internal tools. With a background in UI/UX, clean architecture, and real-world leadership experience, I bring both the code and the clarity to every product I touch.";
   static const String ABOUT_DEV =
-      "Product Designer and Digital Creative Director working in design field for 13 years so far, specialized in UI/UX, Branding and digital designs.";
+      "Mobile Developer with over more than 5 years of experience working in the industry with a knack for solving complex performance problems, refining user experiences, and architecting scalable applications.";
   static const String ABOUT_ME_1 =
       "Product Designer and Digital Creative Director working in design field for";
   static const String ABOUT_ME_2 =
       "13 years so far, specialized in UI/UX, Branding and digital designs.";
   static const String DOWNLOAD_CV = "Download CV";
   static const String EMAIL = "Email";
-  static const String DEV_EMAIL_2 = "nimbus2000@mail.com";
-  static const String BEHANCE = "Behance";
-  static const String BEHANCE_ID = "behance.net/nimbus200";
+  static const String DEV_EMAIL_2 = "me@rohitashwa.co.in";
+  static const String LINKED_IN = "LinkedIn";
+  static const String GITHUB = "Github";
+  static const String LINKED_IN_ID =
+      "https://www.linkedin.com/in/rohitashwa-singh-mobile-apps-flutter-android-ios/";
   static const String HIRE_ME_NOW = "Hire Me Now";
-  static const String UI_UX = "UI/UX Designer";
-  static const String UI_UX_DESC = "creative websites for you";
-  static const String PHOTOGRAPHER = "Photographer";
-  static const String PHOTOGRAPHER_DESC = "Awesome photo & video";
+  static const String JOB_ROLE = "Mobile Developer";
+  static const String MOBILE_DESC = "creative apps for you";
+  static const String DESIGNER = "Designer";
+  static const String DESIGNER_DESC = "Awesome figma designs";
   static const String FREELANCER = "Freelancer";
   static const String FREELANCER_DESC = "You can hire me now";
 
-  static const String PYTHON = "Python";
+  static const String FLUTTER = "Flutter";
   static const String JAVA = "Java";
   static const String C_PLUS_PLUS = "C++";
   static const String JAVASCRIPT = "Javascript";
@@ -54,7 +60,7 @@ class StringConst {
   static const String THERE = "There!";
   static const String ABOUT_ME = "about me";
   static const String ABOUT_ME_DESC =
-      "Product Designer and Digital Creative Director working in design field for 13 years so far, specialized in UI/UX, Branding and digital designs. These are the words we live by in everything we do. Every story we tell, every brand we build, and every interaction we create must not only look beautiful. beautifully, too.";
+      "Mobile developer and entrepreneur with a strong background in building and growing startups. From leading teams to developing innovative mobile solutions, I understand the fast-paced, ever-changing needs of a startup. My experience spans creating user-friendly, efficient apps that help businesses scale while ensuring customer satisfaction. I’ve successfully taken on leadership roles, built cross-functional teams, and driven projects that directly contribute to business growth. Whether it’s improving user experiences, reducing costs, or speeding up product delivery, I bring a results-driven approach that will help a startup thrive. Additionally, I’m skilled in designing intuitive UI/UX for apps and have proficiency in tools like Figma, Canva, and Photoshop to ensure the visual appeal matches the functionality.";
   static const String ABOUT_ME_DESC_1 =
       "Product Designer and Digital Creative Director working in design ";
   static const String ABOUT_ME_DESC_2 =
@@ -65,7 +71,7 @@ class StringConst {
       "story we tell, every brand we build, and every interaction we";
   static const String ABOUT_ME_DESC_5 =
       "create must not only look beautiful. beautifully, too.";
-  static const String CREATIVE_DESIGN = "Need Creative Design?";
+  static const String STARTUP = "Working on a new startup?";
   static const String HELP = "I can Help You!";
   static const String FOLLOW_ME_1 = "You can also follow me here:";
   static const String DRIBBLE = "Dribble";
@@ -73,8 +79,8 @@ class StringConst {
 
   //Skills Section
   static const String MY_SKILLS = "my skills";
-  static const String SKILLS_TITLE_1 = "What My Design";
-  static const String SKILLS_TITLE_2 = "Skills Include";
+  static const String SKILLS_TITLE_1 = "What My SkillSet";
+  static const String SKILLS_TITLE_2 = "Include";
   static const String SKILLS_DESC =
       "These are the words we live by in everything we do. Every story we tell, every brand we build, and every interaction we create must not only look beautiful.";
   static const String TOOLS_1_NAME = "Photoshop";
@@ -83,7 +89,7 @@ class StringConst {
   static const String TOOLS_2_PROFICIENCY = "90%";
   static const String TOOLS_3_NAME = "Sketch";
   static const String TOOLS_3_PROFICIENCY = "70%";
-  static const String SKILLS_1 = "UI/UX Design";
+  static const String SKILLS_1 = "Flutter";
   static const String SKILLS_1_DESC =
       "We believe direct collaboration with our clients is essential to bringing unique";
   static const String SKILLS_2 = "Graphic Design";
@@ -92,9 +98,14 @@ class StringConst {
   static const String SKILLS_3 = "Motion Design";
   static const String SKILLS_3_DESC =
       "We believe direct collaboration with our clients is essential to bringing unique";
-  static const String SKILLS_4 = "Photo Retouch";
+  static const String SKILLS_4 = "UI/UX Design";
   static const String SKILLS_4_DESC =
       "We believe direct collaboration with our clients is essential to bringing unique";
+  static const String SKILLS_5 = "Firebase";
+  static const String SKILLS_6 = "Android";
+  static const String SKILLS_7 = "iOS";
+  static const String SKILLS_8 = "NodeJs";
+  static const String SKILLS_9 = "Java";
 
   //Stats Section
   static const String HAPPY_CLIENTS = "Happy Clients";
@@ -110,23 +121,86 @@ class StringConst {
   static const String MEET_MY_PROJECTS = "Meet My Projects";
   static const String MY_WORKS = "my works";
   static const String PROJECTS_DESC =
-      "Product Designer and Digital Creative Director working in design field for 13 years so far, specialized in UI/UX, Branding and digital designs. These are the words we live by in everything we do.";
+      "Mobile App Developer and Entrepreneur working in the industry for 05 years so far, specialized in product development, delivering the best possible solutions so far. These are the words we live by in everything we do.";
   static const String ALL_PROJECTS = "All Projects";
   static const String ALL = "All";
-  static const String BRANDING = "Branding";
-  static const String PACKAGING = "Packaging";
+  static const String WIOM = "Wiom";
+  static const String REAL_SOL = "Real Solutions";
+  static const String NEO_DOCS = "Neodocs";
+  static const String GMAGICA = "Gmagica";
+  static const String PERSONAL_PROJECTS = "Personal Projects";
   static const String PHOTOGRAPHY = "Photography";
   static const String WEB_DESIGN = "Web Design";
   static const String PORTFOLIO_1_TITLE = "Urban Experience";
   static const String PORTFOLIO_2_TITLE = "Graphic Design";
-  static const String PORTFOLIO_3_TITLE = "Launch Product";
+  static const String PARTNER_WIOM_V1 = "Wiom Partner V1";
+  static const String PARTNER_WIOM_V1_LINK =
+      "https://play.google.com/store/apps/details?id=com.i2e1.wiom.sales&hl=en_IN";
+  static const String PARTNER_WIOM_V2 = "Wiom Partner V2";
+  static const String PARTNER_WIOM_V2_LINK =
+      "https://wiom-partner.s3.ap-south-1.amazonaws.com/Wiom-Partner.apk";
+  static const String EXPERT_WIOM = "Wiom Expert V2";
+  static const String EXPERT_WIOM_LINK =
+      "https://wiom-partner.s3.ap-south-1.amazonaws.com/Wiom-Expert.apk";
+  static const String CUSTOMER_WIOM = "Wiom Customer";
+  static const String CUSTOMER_WIOM_LINK =
+      "https://play.google.com/store/apps/details?id=com.i2e1.wiom_gold&hl=en_IN";
+  static const String LMS_WIOM = "WIOM LMS";
+  static const String LMS_WIOM_LINK =
+      "https://play.google.com/store/apps/details?id=com.i2e1.wiom.leads&hl=en_IN";
+  static const String PROGRAMMING_HUB_REAL_SOL =
+      "Programming Hub :\nLearn to Code";
+  static const String PROGRAMMING_HUB_REAL_SOL_LINK =
+      "https://play.google.com/store/apps/details?id=com.freeit.java&hl=en_IN";
+  static const String GRATEFUL_DIARY_REAL_SOL =
+      "Grateful Diary :\nDaily Journal";
+  static const String GRATEFUL_DIARY_REAL_SOL_LINK =
+      "https://play.google.com/store/apps/details?id=grateful.gratefulapp.gratitude.happiness.wellbeing.dailyjournal.journal&hl=en_IN";
+  static const String NEODOCS_KIDNEY_TEST = "Neodocs - ACR, GFR Kidney Test";
+  static const String NEODOCS_KIDNEY_TEST_LINK =
+      "https://play.google.com/store/apps/details?id=com.app.neodocs&hl=en_IN";
+  static const String PREGAFAITH = "Pregafaith";
+  static const String PREGAFAITH_LINK =
+      "https://play.google.com/store/apps/details?id=com.neodocs.app.pregnancy.test&hl=en_IN";
+  static const String INSTANT_MALARIA_DENGUE = "Instant Malaria & Dengue Test";
+  static const String INSTANT_MALARIA_DENGUE_LINK =
+      "https://play.google.com/store/apps/details?id=com.neodocs.app.malaria&hl=en_IN";
+  static const String MOOV_MONEY_TOGO = "MOOV MONEY TOGO";
+  static const String MOOV_MONEY_TOGO_LINK =
+      "https://flooz-moov-togo.en.aptoide.com/app";
+  static const String MOOV_MONEY_BENIN = "MOOV MONEY BENIN";
+  static const String MOOV_MONEY_BENIN_LINK =
+      "https://play.google.com/store/apps/details?id=com.tlc.flous.subs.bn&hl=en";
+  static const String MOOV_MONEY_CI = "MOOV MONEY CI";
+  static const String MOOV_MONEY_CI_LINK =
+      "https://play.google.com/store/apps/details?id=ci.moovmoney.mmpayapi&hl=en_IN";
+  static const String JAMMERZ = "Jammerz : The Music Streaming App";
+  static const String JAMMERZ_LINK = "https://practice-projects-7c9bf.web.app/";
   static const String PORTFOLIO_4_TITLE = "3D Animation";
   static const String PORTFOLIO_5_TITLE = "Collect Idea";
   static const String PORTFOLIO_6_TITLE = "Company Branding";
 
+  //HOBBIES & Section
+  static const String MY_HOBBIES = "my hobbies";
+  static const String HOBBIES_TITLE = "HOBBIES";
+  static const String HOBBIES_DESC =
+      "Since a kid I have been keenly exploring and have experienced a lot of hobbies";
+  static const String HOBBIES_TYPE_TITLE_1 = "PROFESSIONAL";
+  static const String HOBBIES_TYPE_TITLE_2 = "NOT SO PROFESSIONAL";
+  static const String HOBBY_1 = "Product Development";
+  static const String HOBBY_2 = "Photography";
+  static const String HOBBY_3 = "Designing";
+  static const String HOBBY_4 = "Entrepreneur";
+  static const String HOBBY_5 = "Copy writing";
+  static const String HOBBY_6 = "Travelling";
+  static const String HOBBY_7 = "Bike Riding";
+  static const String HOBBY_8 = "Gym";
+  static const String HOBBY_9 = "Football";
+  static const String HOBBY_10 = "Film Making";
+
   //Awards & Section
   static const String MY = "My";
-  static const String CV = "CV";
+  static const String CV = "RESUME";
   static const String MY_AWARDS = "my awards";
   static const String AWARDS_TITLE = "Honours & Awards";
   static const String AWARDS_DESC =
@@ -180,16 +254,16 @@ class StringConst {
   //Footer Section
   static const String LETS_TALK = "Let's Talk!";
   static const String PHONE_ME = "Phone Me";
-  static const String PHONE_NUMBER = "+1 800 123 456 789";
+  static const String PHONE_NUMBER = "+91 9354859197";
   static const String MAIL_ME = "Mail Me";
   static const String FOLLOW_ME_2 = "Follow Me";
   static const String FOLLOW_ME_2_URL = "behance.net/nimbus2000";
   static const String RIGHTS_RESERVED = "© 2021 All rights reserved";
   static const String BUILT_BY = "Built with Flutter by";
-  static const String DAVID_COBBINA = "David Cobbina";
+  static const String ROHITASHWA_SINGH = "Rohitashwa Singh";
   static const String DESIGNED_BY = "Designed by";
   static const String WEB_GENIUS_LAB = "WebGeniusLab";
-  static const String MADE_IN_GHANA = "Made in";
+  static const String MADE_IN_INDIA = "Made in";
   static const String WITH_LOVE = "with";
   static const String FIND_ME = "Find me on..";
 
@@ -197,9 +271,10 @@ class StringConst {
   static const String WEB_GENIUS_LAB_URL = "https://webgeniuslab.net/";
   static const String BEHANCE_URL = "https://www.behance.net/";
   static const String DRIBBLE_URL = "https://dribbble.com/";
-  static const String LINKED_IN_URL = "https:linkedin.com/in/";
+  static const String LINKED_IN_URL =
+      "https://www.linkedin.com/in/rohitashwa-singh-mobile-apps-flutter-android-ios/";
   static const String INSTAGRAM_URL = "https://www.instagram.com/";
-  static const String GITHUB_URL = "https://github.com/";
+  static const String GITHUB_URL = "https://github.com/Jokerohit4";
   static const String FACEBOOK_URL = "https://facebook.com/";
   static const String TWITTER_URL = "http://twitter.com/";
   static const String DAVID_LEGEND_URL = "https://davidcobbina.com/#/";
